Updated the List of Makers for import
Updated the check for a list of makers by looping through every maker and adding to the CPT. One thing to note, if a maker field is empty, we need something to fall back to. The best option I could find is the "Project Name", but if there are two unnamed makers, then we have dups...? Check line 384 for more info.